CA-338201: use a more efficient way of splitting a map
There is a builtin way of splitting a Map: Map.split.
This is more efficient than partitioning a map, because partitioning
needs to walk the entire map, whereas splitting can stop as soon as it
finds a sub-tree that is greater than, and piece together subtrees.

I assume that rounding was applied for scheduled events to have some
form of batching, and to have a more efficient way of picking next item
to schedule (we can skip an entire list of jobs scheduled at second 7, if we are at second 5).
However if we have jobs with various schedules the map can still grow an
require a linear walk.

Signed-off-by: Edwin Török <edvin.torok@citrix.com>